move workgroups to module
